Update/to do's to users

+++ api/controllers/todo.js
@@ -8,22 +8,33 @@ const tokenExtractor = middleware.tokenExtractor
 
 //Handle GET all the todos
 todoRouter.get('/', tokenExtractor, userExtractor, async (request, response, next) => {
-  const allToDos = await Todo.find({})
-  response.json(allToDos)
-  next()
+
+  const user = request.user
+  const allTodos = await Todo
+    .find({ user: user.id })
+
+  return response.json(allTodos)
 })
 
 //Handle POST new todos
 todoRouter.post('/', tokenExtractor, userExtractor, async (request, response) => {
 
   const { name } = request.body
+  const user = request.user
+
   const todo = new Todo({
     name: name,
     isComplete: false,
     date: new Date(),
+    user: user.id
   })
+
   const savedTodo = await todo.save()
-  response.status(201).json(savedTodo)
+
+  user.todos = user.todos.concat(savedTodo.id)
+  await user.save()
+
+  return response.status(201).json(savedTodo)
 
 })
 
@@ -31,19 +42,31 @@ todoRouter.post('/', tokenExtractor, userExtractor, async (request, response) =>
 todoRouter.get('/:id', tokenExtractor, userExtractor, async (request, response) => {
   const todoId = request.params.id
   const todoById = await Todo.findById(todoId)
-  response.json(todoById)
+  return response.json(todoById)
 })
 
 //Handle PUT requests
 todoRouter.put('/:id', tokenExtractor, userExtractor, async (request, response) => {
   const updatedTodo = await Todo.findByIdAndUpdate(request.params.id, request.body, { new: true })
-  response.json(updatedTodo)
+  return response.json(updatedTodo)
 })
 
 // DELETE a todo by id
 todoRouter.delete('/:id', tokenExtractor, userExtractor, async (request, response) => {
-  const deletedTodo = await Todo.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+
+  const selectedTodo = await Todo.findById(request.params.id)
+  const user = request.user
+
+  const todoInUser = user?.todos.filter(todo => todo.toString() === selectedTodo?.id)
+
+  if (todoInUser?.toString() === selectedTodo?.id) {
+    await Todo.findByIdAndRemove(request.params.id)
+
+    user.todos = user.todos.filter(todo => todo.toString() !== selectedTodo?.id)
+    await user.save()
+    return response.status(204).end()
+  }
+  return response.status(401).end()
 })
 
 module.exports = todoRouter
